% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{align_max_shift}
\alias{align_max_shift}
\alias{align_segment_length}
\title{Parameters for alignment steps}
\usage{
align_max_shift(range = c(1L, 50L), trans = NULL)

align_segment_length(range = c(10L, 100L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\value{
A function with classes \code{"quant_param"} and \code{"param"}.
}
\description{
\code{align_max_shift()} controls the maximum shift allowed in alignment.
\code{align_segment_length()} controls segment size for COW alignment.
}
\examples{
align_max_shift()
align_segment_length()
}
