% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix-effects.R
\name{autoplot.measure_matrix_effect}
\alias{autoplot.measure_matrix_effect}
\title{Plot Matrix Effects}
\usage{
\method{autoplot}{measure_matrix_effect}(object, type = c("bar", "point", "forest"), show_limits = TRUE, ...)
}
\arguments{
\item{object}{A \code{measure_matrix_effect} object.}

\item{type}{Plot type: "bar", "point", or "forest". Default "bar".}

\item{show_limits}{Show acceptable limits (80-120\%)? Default TRUE.}

\item{...}{Additional arguments (unused).}
}
\value{
A ggplot object.
}
\description{
Creates a visualization of matrix effects showing suppression/enhancement.
}
