% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-builtin.R
\name{bigaussian_peak_model}
\alias{bigaussian_peak_model}
\title{Create Bi-Gaussian Peak Model}
\usage{
bigaussian_peak_model()
}
\value{
A \code{bigaussian_peak_model} object.
}
\description{
Creates a Bi-Gaussian peak model with four parameters:
height, center, width_left, and width_right.
}
\details{
The Bi-Gaussian function uses different widths on the left and right
sides of the peak, providing flexible asymmetry.
}
\examples{
model <- bigaussian_peak_model()
x <- seq(0, 10, by = 0.1)
params <- list(height = 1, center = 5, width_left = 0.8, width_right = 1.2)
y <- peak_model_value(model, x, params)
plot(x, y, type = "l")

}
\seealso{
Other peak-models: 
\code{\link{emg_peak_model}()},
\code{\link{gaussian_peak_model}()},
\code{\link{lorentzian_peak_model}()}
}
\concept{peak-models}
