% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{bin_width}
\alias{bin_width}
\alias{emsc_degree}
\alias{osc_n_components}
\title{Parameters for feature engineering and scatter correction}
\usage{
bin_width(range = c(1, 20), trans = NULL)

emsc_degree(range = c(0L, 4L), trans = NULL)

osc_n_components(range = c(1L, 10L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\value{
A function with classes \code{"quant_param"} and \code{"param"}.
}
\description{
\code{bin_width()} controls the width of bins in spectral binning.
\code{emsc_degree()} controls the polynomial degree for EMSC correction.
\code{osc_n_components()} controls the number of orthogonal components in OSC.
}
\examples{
bin_width()
emsc_degree()
osc_n_components()
}
