% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{button_plot}
\alias{button_plot}
\title{button_plot Function}
\usage{
button_plot(min_x, max_x, y, text, back_col, text_col)
}
\arguments{
\item{min_x}{minimum x coord}

\item{max_x}{max x coord}

\item{y}{y cords of box}

\item{text}{text in box}

\item{back_col}{background colour}

\item{text_col}{text colour}
}
\description{
Plots buttons on image
}
