% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{clickPosition}
\alias{clickPosition}
\title{insideFunc Function}
\usage{
clickPosition(coords, id)
}
\arguments{
\item{coords}{coords to check}

\item{id}{Image details}
}
\description{
Are coords within button borders?
}
