% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_units.R
\name{convert_units}
\alias{convert_units}
\title{Convert Units}
\usage{
convert_units(x, units)
}
\arguments{
\item{x}{a \code{\link{measure}} object}

\item{units}{a character value}
}
\value{
A \code{measure} object with new units
}
\description{
Convert the units of a \code{measure} object to newly specified units.
}
