% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-class.R
\name{create_peak_model}
\alias{create_peak_model}
\title{Create a Peak Model by Name}
\usage{
create_peak_model(name)
}
\arguments{
\item{name}{Name of the model (e.g., "gaussian", "emg", "bigaussian").}
}
\value{
A \code{peak_model} object.
}
\description{
Creates a peak model object from a registered model name.
}
\examples{
model <- create_peak_model("gaussian")
print(model)

}
\seealso{
\code{\link[=peak_models]{peak_models()}}, \code{\link[=register_peak_model]{register_peak_model()}}
}
