% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{derivative_order}
\alias{derivative_order}
\alias{derivative_gap}
\alias{derivative_segment}
\title{Parameters for derivative steps}
\usage{
derivative_order(range = c(1L, 2L), trans = NULL)

derivative_gap(range = c(1L, 10L), trans = NULL)

derivative_segment(range = c(1L, 5L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\value{
A function with classes \code{"quant_param"} and \code{"param"}.
}
\description{
\code{derivative_order()} controls the order of differentiation in
\code{step_measure_derivative()} (1 = first derivative, 2 = second derivative).
\code{derivative_gap()} and \code{derivative_segment()} control the gap derivative
(Norris-Williams) parameters in \code{step_measure_derivative_gap()}.
}
\examples{
derivative_order()
derivative_gap()
derivative_segment()
}
