% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostics and Visualization for Measure Data}
\description{
Autoplot and fortify methods for visualizing spectral/chromatographic data
and recipe transformations.
}
\keyword{internal}
