% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-builtin.R
\name{emg_peak_model}
\alias{emg_peak_model}
\title{Create EMG Peak Model}
\usage{
emg_peak_model()
}
\value{
An \code{emg_peak_model} object.
}
\description{
Creates an Exponentially Modified Gaussian peak model with four parameters:
height, center, width (sigma), and tau (exponential decay constant).
}
\details{
The EMG function models asymmetric peaks with tailing, common in
chromatography. It is the convolution of a Gaussian with an exponential
decay function.
}
\examples{
model <- emg_peak_model()
x <- seq(0, 15, by = 0.1)
params <- list(height = 1, center = 5, width = 0.5, tau = 0.3)
y <- peak_model_value(model, x, params)
plot(x, y, type = "l")

}
\seealso{
Other peak-models: 
\code{\link{bigaussian_peak_model}()},
\code{\link{gaussian_peak_model}()},
\code{\link{lorentzian_peak_model}()}
}
\concept{peak-models}
