% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2-scales.R
\name{enmeasure_scale}
\alias{enmeasure_scale}
\title{Add Units to A Scale}
\usage{
enmeasure_scale(super)
}
\arguments{
\item{super}{a \code{\link[ggplot2:ggplot2-ggproto]{ggplot2::Scale}} object returned from a \verb{scale_*()} function
like \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}}
}
\value{
A \code{\link[ggplot2:ggplot2-ggproto]{ggplot2::Scale}} object
}
\description{
Create a \code{\link[ggplot2:ggproto]{ggplot2::ggproto}} object that inherits from \code{super}, designed
to be a \code{ggplot2} scale for \code{measure} objects that adds units to the
aesthetic label.
}
\details{
\code{make_title()} is redefined to return a formula derived from
\code{super$make_title()} and \code{units}. See \code{\link[=plot_math]{plot_math()}}.

\code{transform()} is redefined to set \code{units} to \code{units(x)} and pass a plain
\code{double} to \code{super$transform()}.
}
\examples{
library(ggplot2)
p <- ggplot(mpg, aes(measure(displ, "L"))) + geom_histogram()
s <- scale_x_continuous("Engine displacement")
p + enmeasure_scale(s)

}
\seealso{
\code{\link[=scale_x_measure]{scale_x_measure()}}
}
