% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_area_functions.R
\name{extract_area}
\alias{extract_area}
\title{extract_length Function}
\usage{
extract_area(data)
}
\arguments{
\item{data}{}
}
\description{
Extract total area inside plotted points in pixels. Uses get_outline to get the outline, then works out how many pixels are inside it, include the outline
}
\examples{
#extract_length()
}
