% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_area_functions.R
\name{extract_length}
\alias{extract_length}
\title{extract_length Function}
\usage{
extract_length(data)
}
\arguments{
\item{data}{}
}
\description{
Extract total length of plotted line in pixels
}
\examples{
#extract_length()
}
