% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_area_functions.R
\name{extract_length_image}
\alias{extract_length_image}
\title{extract_length_image Function}
\usage{
extract_length_image(data, metric = "pixels", pixels_per = NULL)
}
\arguments{
\item{data}{}

\item{pixels_per}{}
}
\description{
Extract total length of plotted line in pixels
}
