% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-class.R
\name{find_measure_cols}
\alias{find_measure_cols}
\title{Find measure columns in a data frame}
\usage{
find_measure_cols(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Character vector of column names containing measure data.
Returns empty character vector if no measure columns found.
}
\description{
Finds all columns in a data frame that contain measurement data
(i.e., are of class \code{measure_list}).
}
\examples{
library(recipes)

rec <- recipe(water + fat + protein ~ ., data = meats_long) |>
  update_role(id, new_role = "id") |>
  step_measure_input_long(transmittance, location = vars(channel)) |>
  prep()

result <- bake(rec, new_data = NULL)
find_measure_cols(result)  # ".measures"

}
