% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_validate.R
\name{find_measure_nd_cols}
\alias{find_measure_nd_cols}
\title{Find n-dimensional measure columns in a data frame}
\usage{
find_measure_nd_cols(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Character vector of column names.
}
\description{
Returns the names of columns that contain \code{measure_nd_list} objects.
}
\examples{
# After using step_measure_input_long with multiple location columns
# find_measure_nd_cols(result)

}
