% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-operations.R
\name{find_peaks_cols}
\alias{find_peaks_cols}
\title{Find peaks columns in a data frame}
\usage{
find_peaks_cols(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
Character vector of column names.
}
\description{
Find peaks columns in a data frame
}
