% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-class.R
\name{fortify.measure_calibration}
\alias{fortify.measure_calibration}
\title{Extract Calibration Curve Data}
\usage{
\method{fortify}{measure_calibration}(model, data = NULL, ...)
}
\arguments{
\item{model}{A \link{measure_calibration} object.}

\item{data}{Ignored.}

\item{...}{Additional arguments (unused).}
}
\value{
A data frame with the calibration data and fitted values/residuals.
}
\description{
S3 method to extract the underlying data from a calibration object
in a format suitable for ggplot2.
}
