% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-builtin.R
\name{gaussian_peak_model}
\alias{gaussian_peak_model}
\title{Create Gaussian Peak Model}
\usage{
gaussian_peak_model()
}
\value{
A \code{gaussian_peak_model} object.
}
\description{
Creates a symmetric Gaussian peak model with three parameters:
height, center, and width (sigma).
}
\details{
The Gaussian function is:
\deqn{f(x) = h \cdot \exp\left(-\frac{(x - c)^2}{2\sigma^2}\right)}

where \code{h} is height, \code{c} is center, and \code{sigma} is width.
}
\examples{
model <- gaussian_peak_model()
x <- seq(0, 10, by = 0.1)
params <- list(height = 1, center = 5, width = 1)
y <- peak_model_value(model, x, params)
plot(x, y, type = "l")

}
\seealso{
Other peak-models: 
\code{\link{bigaussian_peak_model}()},
\code{\link{emg_peak_model}()},
\code{\link{lorentzian_peak_model}()}
}
\concept{peak-models}
