% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-class.R
\name{get_measure_col_ndim}
\alias{get_measure_col_ndim}
\title{Get the dimensionality of a measure column}
\usage{
get_measure_col_ndim(data, col)
}
\arguments{
\item{data}{A data frame.}

\item{col}{Character string naming the measure column.}
}
\value{
Integer indicating the number of dimensions.
}
\description{
Returns the number of dimensions (1 for \code{measure_list}, 2+ for
\code{measure_nd_list}) of a measure column in a data frame.
}
\examples{
library(recipes)

rec <- recipe(water + fat + protein ~ ., data = meats_long) |>
  update_role(id, new_role = "id") |>
  step_measure_input_long(transmittance, location = vars(channel)) |>
  prep()

result <- bake(rec, new_data = NULL)
get_measure_col_ndim(result, ".measures")  # 1

}
