% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_area_functions.R
\name{get_outline}
\alias{get_outline}
\title{get_outline Function}
\usage{
get_outline(group_data)
}
\arguments{
\item{group_data}{}
}
\description{
Impute outline from plotted points. connects sequentially plotted points with straight line, then finds max and min for each pixel, to get outline
}
