% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-glucose_bioreactors.R
\docType{data}
\name{glucose_bioreactors}
\alias{glucose_bioreactors}
\alias{bioreactors_large}
\alias{bioreactors_small}
\title{Raman Spectra Bioreactor Data}
\source{
Kuhn and Johnson (2020), \emph{Feature Engineering and Selection},
Chapman and Hall/CRC . \url{https://bookdown.org/max/FES/} and
\url{https://github.com/topepo/FES}
}
\value{
Two tibbles. For each, there are 2,651 columns whose names are
numbers and these are the measured assay values (and the names are the wave
numbers). The numeric column \code{glucose} has the outcome data, \code{day} is the
number of days in the bioreactor, the \code{batch_id} is the reactor identifier
(with "L" for large and "S" for small), and \code{batch_sample} that is the ID
and the day.
}
\description{
Kuhn and Johnson (2013) used these two data sets to model the glucose yeild
in large- and small-scale bioreactors:
}
\details{
\itemize{
\item Fifteen small-scale (5 liters) bioreactors were seeded with cells and
were monitored daily for 14 days.
\item Three large-scale bioreactors were also seeded with cells from the same
batch and monitored daily for 14 days.
}

Samples were collected each day from all bioreactors and glucose was
measured. The goal would be to create models on the data from the more
numerous small-scale bioreactors and then evaluate if these results can
accurately predict what is happening in the large-scale bioreactors.
}
\examples{

data(glucose_bioreactors)
dim(bioreactors_small)
}
\keyword{datasets}
