% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-class.R
\name{has_peak_model}
\alias{has_peak_model}
\title{Check if a Peak Model Exists}
\usage{
has_peak_model(name)
}
\arguments{
\item{name}{Model name.}
}
\value{
Logical \code{TRUE} if model exists.
}
\description{
Check if a Peak Model Exists
}
