% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-package.R
\name{hidden_aliases}
\alias{hidden_aliases}
\alias{\%\%,ANY,measure-method}
\alias{\%\%,measure,ANY-method}
\alias{\%\%,measure,measure-method}
\alias{\%/\%,ANY,measure-method}
\alias{\%/\%,measure,ANY-method}
\alias{\%/\%,measure,measure-method}
\alias{*,ANY,measure-method}
\alias{*,measure,ANY-method}
\alias{*,measure,measure-method}
\alias{+,ANY,measure-method}
\alias{+,measure,ANY-method}
\alias{+,measure,measure-method}
\alias{+,measure,missing-method}
\alias{-,ANY,measure-method}
\alias{-,measure,ANY-method}
\alias{-,measure,measure-method}
\alias{-,measure,missing-method}
\alias{/,ANY,measure-method}
\alias{/,measure,ANY-method}
\alias{/,measure,measure-method}
\alias{Compare,ANY,measure-method}
\alias{Compare,measure,ANY-method}
\alias{Compare,measure,measure-method}
\alias{Summary,measure-method}
\alias{[,measure-method}
\alias{[<-,measure,ANY,ANY,ANY-method}
\alias{[[,measure-method}
\alias{[[<-,measure,ANY,ANY,ANY-method}
\alias{^,ANY,measure-method}
\alias{^,measure,ANY-method}
\alias{^,measure,measure-method}
\alias{^,measure,numeric-method}
\alias{as.difftime,measure-method}
\alias{c,measure-method}
\alias{length<-,measure-method}
\alias{plot,measure,measure-method}
\alias{plot,ANY,measure-method}
\alias{plot,measure,ANY-method}
\alias{rep,measure-method}
\alias{rep_len,measure-method}
\alias{rep.int,measure-method}
\alias{show,measure-method}
\title{For S4 methods that require a documentation entry but only clutter the index.}
\description{
For S4 methods that require a documentation entry but only clutter the index.
}
\keyword{internal}
