% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{image_details}
\alias{image_details}
\title{image_details Function}
\usage{
image_details(image, xlim = NULL, ylim = NULL)
}
\arguments{
\item{image}{path to image}

\item{xlim}{subset of image to plot. Default plots all x}

\item{ylim}{subset of image to plot. Default plots all y}
}
\description{
Plots image
}
