% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{insideFunc}
\alias{insideFunc}
\title{insideFunc Function}
\usage{
insideFunc(coords, x, y)
}
\arguments{
\item{coords}{coords to check}

\item{x}{x box coords}

\item{y}{y box coords}
}
\description{
Are coords within button borders?
}
