% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_measure.R
\name{is_measure}
\alias{is_measure}
\alias{is.measure}
\alias{as_measure}
\alias{as.measure}
\title{Check for and Coerce to Measure}
\usage{
is_measure(x)

is.measure(x)

as_measure(x)

as.measure(x)
}
\arguments{
\item{x}{an R object}
}
\value{
For \code{is_measure()}, a logical value.
For \code{as_measure()}, the result of \code{\link[=measure]{measure()}}
}
\description{
Check if an object inherits the \code{measure} class, or coerce to \code{measure}.
}
\details{
\code{is.measure()} and \code{as.measure()} are aliases of \code{is_measure()} and
\code{as_measure()} respectively.
}
\examples{
(x <- measure(1:5, "m"))
is_measure(x)
(y <- as_measure(1:5))
is_measure(y)

}
