% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration-class.R
\name{is_measure_calibration}
\alias{is_measure_calibration}
\title{Test if Object is a Calibration Curve}
\usage{
is_measure_calibration(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical: TRUE if \code{x} is a \code{measure_calibration} object.
}
\description{
Test if Object is a Calibration Curve
}
\examples{
# After fitting a calibration curve
data <- data.frame(
  nominal_conc = c(0, 10, 25, 50, 100),
  response = c(0.5, 15.2, 35.8, 72.1, 148.3)
)
cal <- measure_calibration_fit(data, response ~ nominal_conc)
is_measure_calibration(cal)

}
