% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-class.R
\name{is_measure_list}
\alias{is_measure_list}
\title{Test if object is a measure list}
\usage{
is_measure_list(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical indicating if \code{x} inherits from \code{measure_list}.
}
\description{
Test if object is a measure list
}
\examples{
# After using step_measure_input_*, the .measures column is a measure_list
library(recipes)

rec <- recipe(water + fat + protein ~ ., data = meats_long) |>
  update_role(id, new_role = "id") |>
  step_measure_input_long(transmittance, location = vars(channel)) |>
  prep()

result <- bake(rec, new_data = NULL)
is_measure_list(result$.measures)

}
