% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_class.R
\name{is_measure_nd_list}
\alias{is_measure_nd_list}
\title{Test if object is an n-dimensional measure list}
\usage{
is_measure_nd_list(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical indicating if \code{x} inherits from \code{measure_nd_list}.
}
\description{
Test if object is an n-dimensional measure list
}
\examples{
# Create and test a measure_nd_list
meas1 <- new_measure_nd_tbl(
  location_1 = 1:5,
  location_2 = rep(1, 5),
  value = rnorm(5)
)
ml <- new_measure_nd_list(list(meas1))
is_measure_nd_list(ml)  # TRUE

}
