% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-class.R
\name{is_measure_tbl}
\alias{is_measure_tbl}
\title{Test if object is a measure tibble}
\usage{
is_measure_tbl(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical indicating if \code{x} inherits from \code{measure_tbl}.
}
\description{
Test if object is a measure tibble
}
\examples{
# Create a measure tibble
mt <- measure:::new_measure_tbl(location = 1:5, value = rnorm(5))
is_measure_tbl(mt)

# Regular tibbles are not measure tibbles
is_measure_tbl(tibble::tibble(location = 1:5, value = rnorm(5)))

}
