% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-class.R
\name{is_peak_model}
\alias{is_peak_model}
\title{Test if Object is a Peak Model}
\usage{
is_peak_model(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical indicating if \code{x} is a \code{peak_model}.
}
\description{
Test if Object is a Peak Model
}
