% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-operations.R
\name{is_peaks_list}
\alias{is_peaks_list}
\title{Test if object is a peaks list}
\usage{
is_peaks_list(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical.
}
\description{
Test if object is a peaks list
}
