% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/length_area_functions.R
\name{line_distance}
\alias{line_distance}
\title{line_distance Function}
\usage{
line_distance(x, y)
}
\arguments{
\item{x}{vector of x coordinated}

\item{y}{vector of y coordinated}
}
\description{
Function to measure length
}
\examples{
#line_distance()
}
