% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measuRe.R
\name{measuRe}
\alias{measuRe}
\title{measuRe Function}
\usage{
measuRe(image_folder, x11 = FALSE)
}
\arguments{
\item{image_folder}{Directory path with images inside. Should only contain images to}

\item{x11}{Logical. use x11? defaults to FALSE}
}
\description{
This function allows to measure objects on images in pixels
}
\examples{
#measuRe()
}
