% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_collect.R
\name{measure_collect}
\alias{measure_collect}
\title{Subtract baseline using robust fitting method}
\usage{
measure_collect(data, shape, measures, identifiers)
}
\arguments{
\item{data}{A dataframe containing the variable for baseline subtraction}

\item{shape}{The starting shape of the data, either "long" or "wide" using
tidyr nomenclature}

\item{measures}{Measurement columns in data}

\item{identifiers}{Identifiers in data, a combination of sample identifiers
and experimental variables}
}
\value{
A dataframe
}
\description{
Subtract baseline using robust fitting method
}
