% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_class.R
\name{measure_dim_units}
\alias{measure_dim_units}
\title{Get dimension units of an n-dimensional measurement}
\usage{
measure_dim_units(x)
}
\arguments{
\item{x}{A \code{measure_nd_tbl} or \code{measure_nd_list} object.}
}
\value{
Character vector of dimension units, or \code{NULL} if not set.
}
\description{
Returns the units for each dimension (e.g., "nm", "min").
}
\examples{
m2d <- new_measure_nd_tbl(
  location_1 = 1:10,
  location_2 = rep(1:2, each = 5),
  value = rnorm(10),
  dim_units = c("min", "nm")
)
measure_dim_units(m2d)

}
