% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nd_class.R
\name{measure_ndim}
\alias{measure_ndim}
\title{Get the number of dimensions of a measurement}
\usage{
measure_ndim(x)
}
\arguments{
\item{x}{A \code{measure_tbl}, \code{measure_nd_tbl}, \code{measure_list}, or
\code{measure_nd_list} object.}
}
\value{
Integer indicating the number of dimensions.
}
\description{
Returns the dimensionality of a measurement object. For 1D measurements
(\code{measure_tbl}), returns 1. For n-dimensional measurements
(\code{measure_nd_tbl}), returns the number of location dimensions.
}
\examples{
# 1D measurement
m1d <- new_measure_tbl(location = 1:10, value = rnorm(10))
measure_ndim(m1d)  # 1

# 2D measurement
m2d <- new_measure_nd_tbl(
  location_1 = rep(1:5, each = 3),
  location_2 = rep(1:3, times = 5),
  value = rnorm(15)
)
measure_ndim(m2d)  # 2

}
