% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata-validation.R
\docType{data}
\name{measure_sample_types}
\alias{measure_sample_types}
\title{Canonical Sample Types}
\format{
An object of class \code{character} of length 5.
}
\usage{
measure_sample_types
}
\description{
The allowed values for the \code{sample_type} column in analytical workflows.
}
\keyword{datasets}
