% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-class.R
\name{new_measure_list}
\alias{new_measure_list}
\title{Create a new measure list}
\usage{
new_measure_list(x = list())
}
\arguments{
\item{x}{A list of \code{measure_tbl} objects or tibbles with \code{location}
and \code{value} columns.}
}
\value{
A list with class \code{measure_list}.
}
\description{
Constructor for creating a collection of measurements suitable
for use as a list column in a data frame. Each element should be
a \code{measure_tbl} or tibble with \code{location} and \code{value} columns.
}
\examples{
# Create individual spectra
spec1 <- new_measure_tbl(location = 1:10, value = rnorm(10))
spec2 <- new_measure_tbl(location = 1:10, value = rnorm(10))

# Combine into a measure_list
specs <- new_measure_list(list(spec1, spec2))
specs

}
\seealso{
\code{\link[=new_measure_tbl]{new_measure_tbl()}} for creating individual measurements,
\code{\link[=is_measure_list]{is_measure_list()}} for checking object class.
}
