% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{outlier_threshold}
\alias{outlier_threshold}
\title{Parameters for quality control steps}
\usage{
outlier_threshold(range = c(2, 5), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\value{
A function with classes \code{"quant_param"} and \code{"param"}.
}
\description{
\code{outlier_threshold()} controls the threshold for outlier detection
(in standard deviation or Mahalanobis distance units).
}
\examples{
outlier_threshold()
}
