% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{peak_location_min}
\alias{peak_location_min}
\alias{peak_location_max}
\title{Parameters for peak normalization}
\usage{
peak_location_min(range = c(0, 100), trans = NULL)

peak_location_max(range = c(0, 100), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\value{
A function with classes \code{"quant_param"} and \code{"param"}.
}
\description{
\code{peak_location_min()} and \code{peak_location_max()} define the bounds for
the reference region in peak normalization. These should be specified
in the same units as the location values in your measurement data.
}
\examples{
peak_location_min()
peak_location_max()
}
