% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-class.R
\name{peak_model_initial_guess}
\alias{peak_model_initial_guess}
\title{Generate Initial Parameter Guess}
\usage{
peak_model_initial_guess(model, x, y, peak_idx)
}
\arguments{
\item{model}{A \code{peak_model} object.}

\item{x}{Numeric vector of x values.}

\item{y}{Numeric vector of y values (signal intensity).}

\item{peak_idx}{Integer index of the peak maximum in \code{x} and \code{y}.}
}
\value{
Named list of initial parameter values.
}
\description{
Estimates initial parameter values from the data, providing a starting
point for optimization.
}
\details{
A good initial guess is crucial for successful optimization. The method
should estimate parameters from local features of the data (peak height,
width at half maximum, asymmetry, etc.).
}
\examples{
model <- create_peak_model("gaussian")
x <- seq(0, 10, by = 0.1)
y <- dnorm(x, mean = 5, sd = 1)
peak_idx <- which.max(y)
initial <- peak_model_initial_guess(model, x, y, peak_idx)
initial

}
\seealso{
\code{\link[=peak_model_bounds]{peak_model_bounds()}}
}
