% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-class.R
\name{peak_model_param_names}
\alias{peak_model_param_names}
\title{Get Parameter Names from Peak Model}
\usage{
peak_model_param_names(model)
}
\arguments{
\item{model}{A \code{peak_model} object.}
}
\value{
Character vector of parameter names.
}
\description{
Get Parameter Names from Peak Model
}
