% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_image}
\alias{plot_image}
\title{plot_image Function}
\usage{
plot_image(image, xlim = NULL, ylim = NULL)
}
\arguments{
\item{image}{path to image}

\item{xlim}{subset of image to plot. Default plots all x}

\item{ylim}{subset of image to plot. Default plots all y}
}
\description{
Plots image
}
\examples{
#plot_image()
}
