% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_math.R
\name{plot_math}
\alias{plot_math}
\title{Mathematical Annotation of Measures}
\usage{
plot_math(x, sub = substitute(x))
}
\arguments{
\item{x}{a \code{measure} object or its \code{units()}}

\item{sub}{the main label, passed to \code{\link[=format]{format()}}}
}
\value{
A formula obeying \code{plotmath} syntax rules
}
\description{
Create a formula for \code{\link{plotmath}} plot annotation from a \code{\link{measure}} object.
}
\examples{
x <- measure(1:5, "k|m s^-1")
plot_math(x)
plot(x)
plot(x, ylab = plot_math(x, "Speed"))

}
