% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_points}
\alias{plot_points}
\title{plot_points Function}
\usage{
plot_points(group_data, current_item, current_col = "red", done_col = "blue")
}
\arguments{
\item{group_data}{data to plot from}
}
\description{
Plot previously extracted points
}
