% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-class.R
\name{register_peak_model}
\alias{register_peak_model}
\title{Register a Peak Model}
\usage{
register_peak_model(
  name,
  constructor,
  pack_name,
  description = "",
  technique = NULL
)
}
\arguments{
\item{name}{Model name (e.g., "gaussian", "emg", "fraser_suzuki").}

\item{constructor}{Function that creates the peak model object.}

\item{pack_name}{Source package name.}

\item{description}{Brief description of the model.}

\item{technique}{Optional technique name (e.g., "SEC/GPC").}
}
\value{
Invisible \code{TRUE}.
}
\description{
Registers a peak model constructor with the measure package. Technique packs
can use this to add custom peak shapes.
}
\examples{
\dontrun{
# In a technique pack's R/zzz.R:
register_peak_model(
  name = "fraser_suzuki",
  constructor = fraser_suzuki_model,
  pack_name = pkgname,
  description = "Fraser-Suzuki asymmetric peak",
  technique = "SEC/GPC"
)
}

}
\seealso{
\code{\link[=peak_models]{peak_models()}}, \code{\link[=create_peak_model]{create_peak_model()}}
}
