% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_measure.R
\name{scale_measure}
\alias{scale_measure}
\alias{scale_linewidth_ordinal_measure}
\alias{scale_fill_fermenter_measure}
\alias{scale_linetype_manual_measure}
\alias{scale_color_viridis_b_measure}
\alias{scale_color_viridis_c_measure}
\alias{scale_color_viridis_d_measure}
\alias{scale_fill_ordinal_measure}
\alias{scale_alpha_identity_measure}
\alias{scale_size_datetime_measure}
\alias{scale_linetype_discrete_measure}
\alias{scale_color_discrete_measure}
\alias{scale_shape_continuous_measure}
\alias{scale_shape_binned_measure}
\alias{scale_y_discrete_measure}
\alias{scale_colour_grey_measure}
\alias{scale_colour_gradient2_measure}
\alias{scale_fill_grey_measure}
\alias{scale_fill_distiller_measure}
\alias{scale_colour_identity_measure}
\alias{scale_y_date_measure}
\alias{scale_alpha_measure}
\alias{scale_alpha_ordinal_measure}
\alias{scale_fill_steps_measure}
\alias{scale_alpha_manual_measure}
\alias{scale_alpha_binned_measure}
\alias{scale_size_binned_measure}
\alias{scale_fill_discrete_measure}
\alias{scale_x_datetime_measure}
\alias{scale_y_time_measure}
\alias{scale_color_steps_measure}
\alias{scale_y_continuous_measure}
\alias{scale_colour_viridis_b_measure}
\alias{scale_colour_viridis_c_measure}
\alias{scale_colour_viridis_d_measure}
\alias{scale_size_measure}
\alias{scale_alpha_continuous_measure}
\alias{scale_x_log10_measure}
\alias{scale_color_datetime_measure}
\alias{scale_colour_stepsn_measure}
\alias{scale_color_gradient_measure}
\alias{scale_discrete_manual_measure}
\alias{scale_x_binned_measure}
\alias{scale_fill_steps2_measure}
\alias{scale_linewidth_measure}
\alias{scale_color_gradient2_measure}
\alias{scale_fill_gradient_measure}
\alias{scale_color_grey_measure}
\alias{scale_colour_gradientn_measure}
\alias{scale_size_discrete_measure}
\alias{scale_size_identity_measure}
\alias{scale_color_ordinal_measure}
\alias{scale_alpha_discrete_measure}
\alias{scale_colour_steps_measure}
\alias{scale_alpha_datetime_measure}
\alias{scale_linetype_identity_measure}
\alias{scale_size_continuous_measure}
\alias{scale_color_fermenter_measure}
\alias{scale_fill_continuous_measure}
\alias{scale_x_discrete_measure}
\alias{scale_color_date_measure}
\alias{scale_colour_binned_measure}
\alias{scale_linewidth_continuous_measure}
\alias{scale_fill_identity_measure}
\alias{scale_shape_discrete_measure}
\alias{scale_color_brewer_measure}
\alias{scale_x_sqrt_measure}
\alias{scale_size_manual_measure}
\alias{scale_size_date_measure}
\alias{scale_colour_gradient_measure}
\alias{scale_linewidth_manual_measure}
\alias{scale_fill_binned_measure}
\alias{scale_linewidth_identity_measure}
\alias{scale_size_ordinal_measure}
\alias{scale_fill_date_measure}
\alias{scale_color_distiller_measure}
\alias{scale_continuous_identity_measure}
\alias{scale_colour_hue_measure}
\alias{scale_color_manual_measure}
\alias{scale_color_stepsn_measure}
\alias{scale_linetype_continuous_measure}
\alias{scale_color_continuous_measure}
\alias{scale_x_continuous_measure}
\alias{scale_color_binned_measure}
\alias{scale_shape_measure}
\alias{scale_y_log10_measure}
\alias{scale_colour_brewer_measure}
\alias{scale_fill_datetime_measure}
\alias{scale_fill_stepsn_measure}
\alias{scale_colour_discrete_measure}
\alias{scale_fill_viridis_b_measure}
\alias{scale_linewidth_date_measure}
\alias{scale_fill_viridis_c_measure}
\alias{scale_fill_viridis_d_measure}
\alias{scale_shape_ordinal_measure}
\alias{scale_colour_distiller_measure}
\alias{scale_x_reverse_measure}
\alias{scale_colour_ordinal_measure}
\alias{scale_fill_brewer_measure}
\alias{scale_linewidth_binned_measure}
\alias{scale_alpha_date_measure}
\alias{scale_size_area_measure}
\alias{scale_colour_steps2_measure}
\alias{scale_type_measure}
\alias{scale_colour_manual_measure}
\alias{scale_color_gradientn_measure}
\alias{scale_fill_hue_measure}
\alias{scale_linetype_binned_measure}
\alias{scale_x_date_measure}
\alias{scale_shape_manual_measure}
\alias{scale_radius_measure}
\alias{scale_color_identity_measure}
\alias{scale_x_time_measure}
\alias{scale_color_steps2_measure}
\alias{scale_fill_manual_measure}
\alias{scale_colour_fermenter_measure}
\alias{scale_linewidth_discrete_measure}
\alias{scale_linewidth_datetime_measure}
\alias{scale_fill_gradientn_measure}
\alias{scale_y_datetime_measure}
\alias{scale_shape_identity_measure}
\alias{scale_color_hue_measure}
\alias{scale_y_sqrt_measure}
\alias{scale_y_binned_measure}
\alias{scale_colour_datetime_measure}
\alias{scale_y_reverse_measure}
\alias{scale_fill_gradient2_measure}
\alias{scale_colour_date_measure}
\alias{scale_linetype_measure}
\alias{scale_size_binned_area_measure}
\alias{scale_discrete_identity_measure}
\alias{scale_colour_continuous_measure}
\alias{scale_x_measure}
\alias{scale_y_measure}
\alias{scale_color_measure}
\alias{scale_colour_measure}
\alias{scale_fill_measure}
\title{Scales for Measure Data}
\usage{
scale_linewidth_ordinal_measure(...)

scale_fill_fermenter_measure(...)

scale_linetype_manual_measure(...)

scale_color_viridis_b_measure(...)

scale_color_viridis_c_measure(...)

scale_color_viridis_d_measure(...)

scale_fill_ordinal_measure(...)

scale_alpha_identity_measure(...)

scale_size_datetime_measure(...)

scale_linetype_discrete_measure(...)

scale_color_discrete_measure(...)

scale_shape_continuous_measure(...)

scale_shape_binned_measure(...)

scale_y_discrete_measure(...)

scale_colour_grey_measure(...)

scale_colour_gradient2_measure(...)

scale_fill_grey_measure(...)

scale_fill_distiller_measure(...)

scale_colour_identity_measure(...)

scale_y_date_measure(...)

scale_alpha_measure(...)

scale_alpha_ordinal_measure(...)

scale_fill_steps_measure(...)

scale_alpha_manual_measure(...)

scale_alpha_binned_measure(...)

scale_size_binned_measure(...)

scale_fill_discrete_measure(...)

scale_x_datetime_measure(...)

scale_y_time_measure(...)

scale_color_steps_measure(...)

scale_y_continuous_measure(...)

scale_colour_viridis_b_measure(...)

scale_colour_viridis_c_measure(...)

scale_colour_viridis_d_measure(...)

scale_size_measure(...)

scale_alpha_continuous_measure(...)

scale_x_log10_measure(...)

scale_color_datetime_measure(...)

scale_colour_stepsn_measure(...)

scale_color_gradient_measure(...)

scale_discrete_manual_measure(...)

scale_x_binned_measure(...)

scale_fill_steps2_measure(...)

scale_linewidth_measure(...)

scale_color_gradient2_measure(...)

scale_fill_gradient_measure(...)

scale_color_grey_measure(...)

scale_colour_gradientn_measure(...)

scale_size_discrete_measure(...)

scale_size_identity_measure(...)

scale_color_ordinal_measure(...)

scale_alpha_discrete_measure(...)

scale_colour_steps_measure(...)

scale_alpha_datetime_measure(...)

scale_linetype_identity_measure(...)

scale_size_continuous_measure(...)

scale_color_fermenter_measure(...)

scale_fill_continuous_measure(...)

scale_x_discrete_measure(...)

scale_color_date_measure(...)

scale_colour_binned_measure(...)

scale_linewidth_continuous_measure(...)

scale_fill_identity_measure(...)

scale_shape_discrete_measure(...)

scale_color_brewer_measure(...)

scale_x_sqrt_measure(...)

scale_size_manual_measure(...)

scale_size_date_measure(...)

scale_colour_gradient_measure(...)

scale_linewidth_manual_measure(...)

scale_fill_binned_measure(...)

scale_linewidth_identity_measure(...)

scale_size_ordinal_measure(...)

scale_fill_date_measure(...)

scale_color_distiller_measure(...)

scale_continuous_identity_measure(...)

scale_colour_hue_measure(...)

scale_color_manual_measure(...)

scale_color_stepsn_measure(...)

scale_linetype_continuous_measure(...)

scale_color_continuous_measure(...)

scale_x_continuous_measure(...)

scale_color_binned_measure(...)

scale_shape_measure(...)

scale_y_log10_measure(...)

scale_colour_brewer_measure(...)

scale_fill_datetime_measure(...)

scale_fill_stepsn_measure(...)

scale_colour_discrete_measure(...)

scale_fill_viridis_b_measure(...)

scale_linewidth_date_measure(...)

scale_fill_viridis_c_measure(...)

scale_fill_viridis_d_measure(...)

scale_shape_ordinal_measure(...)

scale_colour_distiller_measure(...)

scale_x_reverse_measure(...)

scale_colour_ordinal_measure(...)

scale_fill_brewer_measure(...)

scale_linewidth_binned_measure(...)

scale_alpha_date_measure(...)

scale_size_area_measure(...)

scale_colour_steps2_measure(...)

scale_type_measure(...)

scale_colour_manual_measure(...)

scale_color_gradientn_measure(...)

scale_fill_hue_measure(...)

scale_linetype_binned_measure(...)

scale_x_date_measure(...)

scale_shape_manual_measure(...)

scale_radius_measure(...)

scale_color_identity_measure(...)

scale_x_time_measure(...)

scale_color_steps2_measure(...)

scale_fill_manual_measure(...)

scale_colour_fermenter_measure(...)

scale_linewidth_discrete_measure(...)

scale_linewidth_datetime_measure(...)

scale_fill_gradientn_measure(...)

scale_y_datetime_measure(...)

scale_shape_identity_measure(...)

scale_color_hue_measure(...)

scale_y_sqrt_measure(...)

scale_y_binned_measure(...)

scale_colour_datetime_measure(...)

scale_y_reverse_measure(...)

scale_fill_gradient2_measure(...)

scale_colour_date_measure(...)

scale_linetype_measure(...)

scale_size_binned_area_measure(...)

scale_discrete_identity_measure(...)

scale_colour_continuous_measure(...)

scale_x_measure(...)

scale_y_measure(...)

scale_color_measure(...)

scale_colour_measure(...)

scale_fill_measure(...)
}
\arguments{
\item{...}{passed to the corresponding \code{ggplot2} scale function of \verb{*_measure()}, such as \code{\link[ggplot2:scale_continuous]{ggplot2::scale_x_continuous()}}}
}
\value{
The result of \code{\link{enmeasure_scale}}\code{(scale_*(...))}
}
\description{
\code{ggplot2} scales that add units from \code{measure} objects to their aesthetic labels
}
