% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{smooth_window}
\alias{smooth_window}
\alias{smooth_sigma}
\alias{fourier_cutoff}
\alias{despike_threshold}
\title{Parameters for smoothing steps}
\usage{
smooth_window(range = c(3L, 21L), trans = NULL)

smooth_sigma(range = c(0.5, 5), trans = NULL)

fourier_cutoff(range = c(0.01, 0.5), trans = NULL)

despike_threshold(range = c(2, 10), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\value{
A function with classes \code{"quant_param"} and \code{"param"}.
}
\description{
\code{smooth_window()} controls the window size for moving average and median
smoothing. \code{smooth_sigma()} controls the standard deviation for Gaussian
smoothing. \code{fourier_cutoff()} controls the frequency cutoff for Fourier
filtering.
}
\examples{
smooth_window()
smooth_sigma()
fourier_cutoff()
}
