% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{step_baseline}
\alias{step_baseline}
\title{Fit and subtract a baseline from a measurment signal}
\usage{
step_baseline(recipe, ..., role = "predictor")
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step.}

\item{role}{Assign the role of new variables.}
}
\description{
Fit and subtract a baseline from a measurment signal
}
