% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_wide.R
\name{step_measure_output_wide}
\alias{step_measure_output_wide}
\title{Reorganize Measurements to Separate Columns}
\usage{
step_measure_output_wide(
  recipe,
  prefix = "measure_",
  role = "predictor",
  trained = FALSE,
  skip = FALSE,
  id = rand_id("measure_output_wide")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{prefix}{A character string used to name the new columns.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[recipes:bake]{bake()}}? While all operations are baked
when \code{\link[recipes:prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\description{
\code{step_measure_output_wide} creates a \emph{specification} of a recipe
step that converts measures to multiple columns (i.e., "wide" format).
}
\details{
This step is designed convert analytical measurements from their internal
data structure to separate columns.
}
