% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savitzky_golay.R
\name{step_measure_savitzky_golay}
\alias{step_measure_savitzky_golay}
\title{Savitzky-Golay Pre-Processing}
\usage{
step_measure_savitzky_golay(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  degree = 3,
  window_size = 11,
  differentiation_order = 0,
  skip = FALSE,
  id = rand_id("measure_savitzky_golay")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose which
variables are affected by the step. See \code{\link[=selections]{selections()}}
for more details. For the \code{tidy} method, these are not
currently used.}

\item{role}{Not used by this step since no new variables are
created. \if{html}{\out{<change if role is used>}}}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{degree}{The polynomial degree to use for smoothing.}

\item{window_size}{The window size to use for smoothing.}

\item{differentiation_order}{An integer for the degree of filtering (zero
indicates no differentiation).}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_measure_savitzky_golay} creates a \emph{specification} of a recipe
step that \if{html}{\out{<what it does>}}
}
\section{Tidying}{
When you \code{\link[=tidy.recipe]{tidy()}} this step, a tibble with columns
\if{html}{\out{<describe tidying>}} is returned.
}

