% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_blank.R
\name{step_subtract_blank}
\alias{step_subtract_blank}
\title{Subtract signal of a "blank" measurement from target signal response}
\usage{
step_subtract_blank(recipe, ..., role = "predictor")
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{...}{One or more selector functions to choose variables
for this step.}

\item{role}{Assign the role of new variables.}
}
\description{
Subtract signal of a "blank" measurement from target signal response
}
