% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{subtract_rf_baseline}
\alias{subtract_rf_baseline}
\title{Subtract baseline using robust fitting method}
\usage{
subtract_rf_baseline(data, yvar, span = 2/3, maxit = c(5, 5))
}
\arguments{
\item{data}{A dataframe containing the variable for baseline subtraction}

\item{yvar}{The name of the column for baseline subtraction}

\item{span}{Controls the amount of smoothing based on the fraction of data
to use in computing each fitted value, defaults to \code{2/3}.}

\item{maxit}{The number of iterations to use the robust fit, defaults to
\code{c(5, 5)} where the first value specifies iterations for asymmetric weighting
function and the second value for symmetric weighting function.}
}
\value{
A dataframe matching column in data plus \code{raw} and \code{baseline} columns
}
\description{
Subtract baseline using robust fitting method
}
\examples{
meats_long \%>\% subtract_rf_baseline(yvar = transmittance)
}
