% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-class.R
\name{sum_peak_models}
\alias{sum_peak_models}
\title{Sum Multiple Peak Models}
\usage{
sum_peak_models(x, models, params_list)
}
\arguments{
\item{x}{Numeric vector of x values.}

\item{models}{List of \code{peak_model} objects (one per peak).}

\item{params_list}{List of parameter lists (one per peak).}
}
\value{
Numeric vector of summed peak values.
}
\description{
Evaluates multiple peak models and sums their contributions.
}
\examples{
# Two overlapping Gaussian peaks
model1 <- create_peak_model("gaussian")
model2 <- create_peak_model("gaussian")
x <- seq(0, 20, by = 0.1)
params1 <- list(height = 1, center = 8, width = 1)
params2 <- list(height = 0.8, center = 12, width = 1.5)
y <- sum_peak_models(x, list(model1, model2), list(params1, params2))
plot(x, y, type = "l")

}
