% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lod-loq.R
\name{tidy.measure_lod}
\alias{tidy.measure_lod}
\title{Tidy LOD/LOQ Results}
\usage{
\method{tidy}{measure_lod}(x, ...)
}
\arguments{
\item{x}{A measure_lod, measure_loq, or measure_lod_loq object.}

\item{...}{Additional arguments (unused).}
}
\value{
A tibble with the limit value(s) and method information.
}
\description{
Tidy LOD/LOQ Results
}
