% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{tunable_measure}
\alias{tunable_measure}
\alias{tunable.step_measure_savitzky_golay}
\title{tunable methods for measure}
\usage{
\method{tunable}{step_measure_savitzky_golay}(x, ...)
}
\arguments{
\item{x}{A recipe step object}

\item{...}{Not used.}
}
\value{
A tibble object.
}
\description{
These functions define what parameters \emph{can} be tuned for specific steps.
They also define the recommended objects from the \code{dials} package that can be
used to generate new parameter values and other characteristics.
}
\keyword{internal}
