% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uncertainty.R
\name{uncertainty_type_a}
\alias{uncertainty_type_a}
\title{Create Type A Uncertainty from Repeated Measurements}
\usage{
uncertainty_type_a(x, name = "Type A", sensitivity = 1)
}
\arguments{
\item{x}{Numeric vector of repeated measurements.}

\item{name}{Name for this uncertainty component.}

\item{sensitivity}{Sensitivity coefficient (default 1).}
}
\value{
An \link{uncertainty_component} object.
}
\description{
Helper function to calculate Type A uncertainty from a vector of
repeated measurements.
}
\examples{
measurements <- c(10.1, 10.3, 9.9, 10.2, 10.0)
u_repeat <- uncertainty_type_a(measurements, "Repeatability")

}
