% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak-model-class.R
\name{validate_peak_model_params}
\alias{validate_peak_model_params}
\title{Validate Peak Model Parameters}
\usage{
validate_peak_model_params(model, params)
}
\arguments{
\item{model}{A \code{peak_model} object.}

\item{params}{Named list of parameters to validate.}
}
\value{
Invisible \code{TRUE} if valid, otherwise throws an error.
}
\description{
Checks that a parameter list has all required parameters for a model.
}
