% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-class.R, R/nd_class.R
\name{print.measure_tbl}
\alias{print.measure_tbl}
\alias{vec_ptype_abbr.measure_tbl}
\alias{print.measure_list}
\alias{vctrs-methods}
\alias{vec_ptype_abbr.measure_list}
\alias{vec_ptype_full.measure_list}
\alias{format.measure_list}
\alias{obj_print_data.measure_list}
\alias{[.measure_list}
\alias{[[.measure_list}
\alias{c.measure_list}
\alias{print.measure_nd_tbl}
\alias{vec_ptype_abbr.measure_nd_tbl}
\alias{print.measure_nd_list}
\alias{vec_ptype_abbr.measure_nd_list}
\alias{vec_ptype_full.measure_nd_list}
\alias{format.measure_nd_list}
\alias{obj_print_data.measure_nd_list}
\alias{[.measure_nd_list}
\alias{[[.measure_nd_list}
\alias{c.measure_nd_list}
\title{vctrs methods for measure classes}
\usage{
\method{print}{measure_tbl}(x, ..., n = NULL, width = NULL)

\method{vec_ptype_abbr}{measure_tbl}(x, ...)

\method{print}{measure_list}(x, ...)

\method{vec_ptype_abbr}{measure_list}(x, ...)

\method{vec_ptype_full}{measure_list}(x, ...)

\method{format}{measure_list}(x, ...)

\method{obj_print_data}{measure_list}(x, ...)

\method{[}{measure_list}(x, i, ...)

\method{[[}{measure_list}(x, i, ...)

\method{c}{measure_list}(...)

\method{print}{measure_nd_tbl}(x, ..., n = NULL, width = NULL)

\method{vec_ptype_abbr}{measure_nd_tbl}(x, ...)

\method{print}{measure_nd_list}(x, ...)

\method{vec_ptype_abbr}{measure_nd_list}(x, ...)

\method{vec_ptype_full}{measure_nd_list}(x, ...)

\method{format}{measure_nd_list}(x, ...)

\method{obj_print_data}{measure_nd_list}(x, ...)

\method{[}{measure_nd_list}(x, i, ...)

\method{[[}{measure_nd_list}(x, i, ...)

\method{c}{measure_nd_list}(...)
}
\arguments{
\item{x}{A measure_list or measure_tbl object.}

\item{...}{Additional arguments (unused).}
}
\description{
These methods enable nice printing of measure objects in tibbles.
}
\keyword{internal}
