% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{window_side}
\alias{window_side}
\alias{differentiation_order}
\title{Parameter for measure steps}
\usage{
window_side(range = c(1L, 5L), trans = NULL)

differentiation_order(range = c(0L, 4L), trans = NULL)
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest and
largest possible values, respectively. If a transformation is specified,
these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{scales::transform_log10()} or \code{scales::transform_reciprocal()}. If not provided,
the default is used which matches the units used in \code{range}. If no
transformation, \code{NULL}.}
}
\value{
A function with classes \code{"quant_param"} and \code{"param"}.
}
\description{
\code{window_side()} and \code{differentiation_order()} are used with Savitzky-Golay
processing.
}
\details{
This parameter is often used to correct for zero-count data in tables or
proportions.
}
\examples{
window_side()
differentiation_order()
}
