% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{basic_plot}
\alias{basic_plot}
\title{basic_plot Function}
\usage{
basic_plot(image, xlim = NULL, ylim = NULL)
}
\arguments{
\item{image}{path to image}

\item{xlim}{subset of image to plot. Default plots all x}

\item{ylim}{subset of image to plot. Default plots all y}
}
\description{
plots image and buttons
}
\examples{
#basic_plot()
}
