% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{button_plot}
\alias{button_plot}
\title{button_plot Function}
\usage{
button_plot(x, y, text, back_col, text_col, border = TRUE, text_cex = 1.2)
}
\arguments{
\item{x}{min and max x coords}

\item{y}{min and max y coords}

\item{text}{text in box}

\item{back_col}{background colour}

\item{text_col}{text colour}
}
\description{
Plots buttons on image
}
