% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure-package.R
\docType{package}
\name{measure-package}
\alias{measure-package}
\title{measure: Automatic Unit Handling for Numeric Vectors}
\description{
Adds a new class 'measure' for which methods of base functions, such as arithmetic and comparative operators and summary functions, are defined such that the units of 'measure' objects are automatically updated. Allows conversion between units and is compatible with various time interval objects.
}
\author{
\strong{Maintainer}: Harry Thompson \email{harry@mayesfield.uk}

}
\keyword{internal}
