% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anylr-package.R
\docType{package}
\name{measure-package}
\alias{measure}
\alias{measure-package}
\title{measure: A Recipes Package to Interface with Tidymodels for Analytical Measurements}
\description{
Analytical measurements...
}
\author{
\strong{Maintainer}: James Wade \email{github@jameshwade.com} (\href{https://orcid.org/0000-0002-9740-1905}{ORCID})

}
\keyword{internal}
