% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.R, R/s4-measure.R
\name{measure}
\alias{measure}
\alias{units,measure-method}
\alias{units<-,measure-method}
\title{Measures}
\usage{
measure(x = double(), units = attr(x, "units"))

\S4method{units}{measure}(x)

\S4method{units}{measure}(x) <- value
}
\arguments{
\item{x}{a numeric vector, or a \code{measure}, \code{\link{difftime}}, \code{\link[lubridate:duration]{lubridate::duration}},
\code{\link[lubridate:interval]{lubridate::interval}}, or \code{\link[lubridate:period]{lubridate::period}} object}

\item{units, value}{a character value. See Details}
}
\value{
An object of class \code{measure}
}
\description{
A class and utility methods for numeric values with units
}
\section{Methods}{


If \code{x} is a \code{lubridate} Duration, Interval or Period object, it will be
converted to \code{difftime} before running the default method.

If \code{x} is already a \code{measure} \code{\link[=convert_units]{convert_units()}} will be called.
}

\section{Units}{

Units are specified in the following format:

\code{"multiplier prefix1|unit1^exponent1 prefix2|unit2^exponent2 ..."}.

The multiplier, prefixes and exponents are optional. For example,
\code{measure(x, "10 c|m^2 s^-1")} will add units of ten centimeters per second,
and \code{measure(x, "m")} simply of meters. Prefixes can be any of the SI prefixes:

\code{Q}, \code{R}, \code{Y}, \code{Z}, \code{E}, \code{P}, \code{T}, \code{G}, \code{M}, \code{k}, \code{h}, \code{da}, \code{d}, \code{c}, \code{m}, \code{µ},
(or \code{micro}), \code{n}, \code{p}, \code{f}, \code{a}, \code{z}, \code{y}, \code{r}, or \code{q}.

Units have no restriction, although \code{\link{difftime}} units are converted to
\code{s}, \code{min}, \code{hour}, \code{day} and \code{week}.
}

\seealso{
\code{\link[=scale_x_measure]{scale_x_measure()}}
}
