% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_buttons}
\alias{plot_buttons}
\title{plot_buttons Function}
\usage{
plot_buttons(id, n)
}
\arguments{
\item{id}{image details}

\item{n}{number of image}
}
\description{
Plots buttons on image
}
