% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot_image}
\alias{plot_image}
\title{plot_image Function}
\usage{
plot_image(image, id)
}
\arguments{
\item{image}{path to image}

\item{id}{image details from image_details()}
}
\description{
Plots image
}
\examples{
#plot_image()
}
