% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R, R/augmentation.R,
%   R/feature-engineering.R, R/mw-calculations.R, R/region-operations.R,
%   R/savitzky_golay.R, R/scatter-extended.R, R/smooth-wavelet.R
\name{required_pkgs.step_measure_align_dtw}
\alias{required_pkgs.step_measure_align_dtw}
\alias{required_pkgs.step_measure_align_ptw}
\alias{required_pkgs.step_measure_augment_noise}
\alias{required_pkgs.step_measure_augment_shift}
\alias{required_pkgs.step_measure_augment_scale}
\alias{required_pkgs.step_measure_integrals}
\alias{required_pkgs.step_measure_ratios}
\alias{required_pkgs.step_measure_moments}
\alias{required_pkgs.step_measure_bin}
\alias{required_pkgs.step_measure_mw_averages}
\alias{required_pkgs.step_measure_mw_fractions}
\alias{required_pkgs.step_measure_mw_distribution}
\alias{required_pkgs.step_measure_resample}
\alias{required_pkgs.recipe}
\alias{required_pkgs.step_measure_savitzky_golay}
\alias{required_pkgs.step_measure_emsc}
\alias{required_pkgs.step_measure_osc}
\alias{required_pkgs.step_measure_smooth_wavelet}
\title{Set package dependencies}
\usage{
\method{required_pkgs}{step_measure_align_dtw}(x, ...)

\method{required_pkgs}{step_measure_align_ptw}(x, ...)

\method{required_pkgs}{step_measure_augment_noise}(x, ...)

\method{required_pkgs}{step_measure_augment_shift}(x, ...)

\method{required_pkgs}{step_measure_augment_scale}(x, ...)

\method{required_pkgs}{step_measure_integrals}(x, ...)

\method{required_pkgs}{step_measure_ratios}(x, ...)

\method{required_pkgs}{step_measure_moments}(x, ...)

\method{required_pkgs}{step_measure_bin}(x, ...)

\method{required_pkgs}{step_measure_mw_averages}(x, ...)

\method{required_pkgs}{step_measure_mw_fractions}(x, ...)

\method{required_pkgs}{step_measure_mw_distribution}(x, ...)

\method{required_pkgs}{step_measure_resample}(x, ...)

\method{required_pkgs}{step_measure_savitzky_golay}(x, ...)

\method{required_pkgs}{step_measure_emsc}(x, ...)

\method{required_pkgs}{step_measure_osc}(x, ...)

\method{required_pkgs}{step_measure_smooth_wavelet}(x, ...)
}
\arguments{
\item{x}{A step object.}

\item{...}{Not used.}
}
\description{
Set package dependencies
}
\keyword{internal}
