% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savitzky_golay.R
\name{required_pkgs.recipe}
\alias{required_pkgs.recipe}
\alias{required_pkgs.step_isomap}
\title{Set package dependencies}
\usage{
\method{required_pkgs}{step_isomap}(x, ...)
}
\arguments{
\item{x}{A step object.}
}
\description{
Set package dependencies
}
