% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savitzky_golay.R
\name{required_pkgs.recipe}
\alias{required_pkgs.recipe}
\alias{required_pkgs.step_measure_savitzky_golay}
\title{Set package dependencies}
\usage{
\method{required_pkgs}{step_measure_savitzky_golay}(x, ...)
}
\arguments{
\item{x}{A step object.}

\item{...}{Not used.}
}
\description{
Set package dependencies
}
