% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savitzky_golay.R
\name{step_measure_savitzky_golay}
\alias{step_measure_savitzky_golay}
\title{Savitzky-Golay Pre-Processing}
\usage{
step_measure_savitzky_golay(
  recipe,
  role = NA,
  trained = FALSE,
  degree = 3,
  window_size = 11,
  differentiation_order = 0,
  skip = FALSE,
  id = rand_id("measure_savitzky_golay")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the
sequence of operations for this recipe.}

\item{role}{Not used by this step since no new variables are
created.}

\item{trained}{A logical to indicate if the quantities for
preprocessing have been estimated.}

\item{degree}{An integer for the polynomial degree to use for smoothing.}

\item{window_size}{An odd integer for the window size to use for smoothing.}

\item{differentiation_order}{An integer for the degree of filtering (zero
indicates no differentiation).}

\item{skip}{A logical. Should the step be skipped when the
recipe is baked by \code{\link[=bake]{bake()}}? While all operations are baked
when \code{\link[=prep]{prep()}} is run, some operations may not be able to be
conducted on new data (e.g. processing the outcome variable(s)).
Care should be taken when using \code{skip = TRUE} as it may affect
the computations for subsequent operations}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_measure_savitzky_golay} creates a \emph{specification} of a recipe
step that smooths and filters the measurement sequence.
}
\details{
This method can both smooth out random noise and reduce between-predictor
correlation. It fits a polynomial to a window of measurements and this results
in fewer measurements than the input. Measurements are assumed to be equally
spaced.

The polynomial degree should be less than the window size. Also, window
size must be greater than polynomial degree. If either case is true, the
original argument values are increased to satisfy these conditions (with a
warning).

\strong{No selectors should be supplied to this step function}. The data should be in
a special internal format produced by \code{\link[=step_measure_input_wide]{step_measure_input_wide()}} or
\code{\link[=step_measure_input_long]{step_measure_input_long()}}.

The measurement locations are reset to integer indices starting at one.
}
\section{Tidying}{
When you \code{\link[=tidy.recipe]{tidy()}} this step, a tibble with columns
\if{html}{\out{<describe tidying>}} is returned.
}

\examples{
if (rlang::is_installed("prospectr")) {
  rec <-
    recipe(water + fat + protein ~ ., data = meats_long) \%>\%
    update_role(id, new_role = "id") \%>\%
    step_measure_input_long(transmittance, location = vars(channel)) \%>\%
    step_measure_savitzky_golay(
      differentiation_order = 1,
      degree = 3,
      window_size = 5
    ) \%>\%
    prep()
}
}
\concept{measure-differencing}
\concept{measure-smoothing}
