% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline.R
\name{subtract_rf_baseline}
\alias{subtract_rf_baseline}
\title{Subtract baseline using robust fitting method}
\usage{
subtract_rf_baseline(data, yvar, span = 2/3, maxit = c(5, 5))
}
\arguments{
\item{data}{A dataframe containing the variable for baseline subtraction}

\item{yvar}{The name of the column for baseline subtraction}

\item{span}{Controls the amount of smoothing based on the fraction of data
to use in computing each fitted value, defaults to \code{2/3}.}

\item{maxit}{The number of iterations to use the robust fit, defaults to
\code{c(5, 5)} where the first value specifies iterations for asymmetric weighting
function and the second value for symmetric weighting function.}
}
\value{
A dataframe matching column in data plus \code{raw} and \code{baseline} columns
}
\description{
A standalone function for robust fitting baseline subtraction using
local regression with iterative reweighting. For use within a recipe
workflow, see \code{\link[=step_measure_baseline_rf]{step_measure_baseline_rf()}}.
}
\examples{
library(dplyr)
meats_long |>
  group_by(id) |>
  subtract_rf_baseline(yvar = transmittance)
}
\seealso{
\code{\link[=step_measure_baseline_rf]{step_measure_baseline_rf()}} for the recipe step version.
}
