% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{tidy.step_pca}
\alias{tidy.step_pca}
\title{Tidy recipe for PCA}
\usage{
\method{tidy}{step_pca}(x, type = "coef", ...)
}
\arguments{
\item{type}{For \code{step_pca}, either \code{"coef"} (for the variable loadings per
component) or \code{"variance"} (how much variance does each component
account for).}
}
\description{
Tidy recipe for PCA
}
