% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_long.R, R/input_wide.R,
%   R/measure_collect.R, R/output_wide.R
\name{tidy.step_measure_input_long}
\alias{tidy.step_measure_input_long}
\alias{tidy.step_measure_input_wide}
\alias{tidy.step_measure_collect}
\alias{tidy.step_measure_output_wide}
\title{Tidy method for step_measure_collect}
\usage{
\method{tidy}{step_measure_input_long}(x, ...)

\method{tidy}{step_measure_input_wide}(x, ...)

\method{tidy}{step_measure_collect}(x, ...)

\method{tidy}{step_measure_output_wide}(x, ...)
}
\arguments{
\item{x}{A recipe object}

\item{...}{Other variables to pass to tidy}
}
\description{
Tidy method for step_measure_collect
}
