% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_long.R, R/input_wide.R, R/output_long.R,
%   R/output_wide.R, R/savitzky_golay.R, R/snv.R
\name{tidy.step_measure_input_long}
\alias{tidy.step_measure_input_long}
\alias{tidy.step_measure_input_wide}
\alias{tidy.step_measure_output_long}
\alias{tidy.step_measure_output_wide}
\alias{tidy.step_measure_savitzky_golay}
\alias{tidy.step_measure_snv}
\title{Tidiers for measure steps}
\usage{
\method{tidy}{step_measure_input_long}(x, ...)

\method{tidy}{step_measure_input_wide}(x, ...)

\method{tidy}{step_measure_output_long}(x, ...)

\method{tidy}{step_measure_output_wide}(x, ...)

\method{tidy}{step_measure_savitzky_golay}(x, ...)

\method{tidy}{step_measure_snv}(x, ...)
}
\arguments{
\item{x}{A step object.}

\item{...}{Not used.}
}
\description{
Tidiers for measure steps
}
\keyword{internal}
