% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_long.R, R/input_wide.R,
%   R/savitzky_golay.R
\name{tidy.step_measure_input_long}
\alias{tidy.step_measure_input_long}
\alias{tidy.step_measure_input_wide}
\alias{tidy.step_measure_savitzky_golay}
\title{Tidiers for measure steps}
\usage{
\method{tidy}{step_measure_input_long}(x, ...)

\method{tidy}{step_measure_input_wide}(x, ...)

\method{tidy}{step_measure_savitzky_golay}(x, ...)
}
\arguments{
\item{x}{A recipe step.}
}
\description{
Tidiers for measure steps
}
