% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_collect.R
\name{tidy.step_measure_collect}
\alias{tidy.step_measure_collect}
\title{Tidy method for step_measure_collect}
\usage{
\method{tidy}{step_measure_collect}(x, ...)
}
\arguments{
\item{x}{A recipe object}

\item{...}{Other variables to pass to tidy}
}
\description{
Tidy method for step_measure_collect
}
