% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baseline-custom.R, R/baseline-py.R,
%   R/parameters.R
\name{tunable.step_measure_baseline_custom}
\alias{tunable.step_measure_baseline_custom}
\alias{tunable.step_measure_baseline_py}
\alias{tunable_measure}
\alias{tunable.step_measure_savitzky_golay}
\alias{tunable.step_measure_baseline_als}
\alias{tunable.step_measure_baseline_poly}
\alias{tunable.step_measure_baseline_rf}
\alias{tunable.step_measure_detrend}
\alias{tunable.step_measure_normalize_peak}
\alias{tunable.step_measure_derivative}
\alias{tunable.step_measure_derivative_gap}
\title{tunable methods for measure}
\usage{
\method{tunable}{step_measure_baseline_custom}(x, ...)

\method{tunable}{step_measure_baseline_py}(x, ...)

\method{tunable}{step_measure_savitzky_golay}(x, ...)

\method{tunable}{step_measure_baseline_als}(x, ...)

\method{tunable}{step_measure_baseline_poly}(x, ...)

\method{tunable}{step_measure_baseline_rf}(x, ...)

\method{tunable}{step_measure_detrend}(x, ...)

\method{tunable}{step_measure_normalize_peak}(x, ...)

\method{tunable}{step_measure_derivative}(x, ...)

\method{tunable}{step_measure_derivative_gap}(x, ...)
}
\arguments{
\item{x}{A recipe step object}

\item{...}{Not used.}
}
\value{
A tibble object.
}
\description{
These functions define what parameters \emph{can} be tuned for specific steps.
They also define the recommended objects from the \code{dials} package that can be
used to generate new parameter values and other characteristics.
}
\keyword{internal}
