% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R, R/baseline-custom.R,
%   R/baseline-extended.R, R/baseline-pls-advanced.R, R/baseline-py.R,
%   R/feature-engineering.R, R/parameters.R, R/scatter-extended.R
\name{tunable.step_measure_align_cow}
\alias{tunable.step_measure_align_cow}
\alias{tunable.step_measure_baseline_custom}
\alias{tunable.step_measure_baseline_airpls}
\alias{tunable.step_measure_baseline_arpls}
\alias{tunable.step_measure_baseline_aspls}
\alias{tunable.step_measure_baseline_iarpls}
\alias{tunable.step_measure_baseline_fastchrom}
\alias{tunable.step_measure_baseline_py}
\alias{tunable.step_measure_bin}
\alias{tunable_measure}
\alias{tunable.step_measure_savitzky_golay}
\alias{tunable.step_measure_baseline_als}
\alias{tunable.step_measure_baseline_poly}
\alias{tunable.step_measure_baseline_rf}
\alias{tunable.step_measure_detrend}
\alias{tunable.step_measure_normalize_peak}
\alias{tunable.step_measure_derivative}
\alias{tunable.step_measure_derivative_gap}
\alias{tunable.step_measure_smooth_ma}
\alias{tunable.step_measure_smooth_median}
\alias{tunable.step_measure_smooth_gaussian}
\alias{tunable.step_measure_filter_fourier}
\alias{tunable.step_measure_despike}
\alias{tunable.step_measure_align_shift}
\alias{tunable.step_measure_qc_outlier}
\alias{tunable.step_measure_emsc}
\alias{tunable.step_measure_osc}
\title{tunable methods for measure}
\usage{
\method{tunable}{step_measure_align_cow}(x, ...)

\method{tunable}{step_measure_baseline_custom}(x, ...)

\method{tunable}{step_measure_baseline_airpls}(x, ...)

\method{tunable}{step_measure_baseline_arpls}(x, ...)

\method{tunable}{step_measure_baseline_aspls}(x, ...)

\method{tunable}{step_measure_baseline_iarpls}(x, ...)

\method{tunable}{step_measure_baseline_fastchrom}(x, ...)

\method{tunable}{step_measure_baseline_py}(x, ...)

\method{tunable}{step_measure_bin}(x, ...)

\method{tunable}{step_measure_savitzky_golay}(x, ...)

\method{tunable}{step_measure_baseline_als}(x, ...)

\method{tunable}{step_measure_baseline_poly}(x, ...)

\method{tunable}{step_measure_baseline_rf}(x, ...)

\method{tunable}{step_measure_detrend}(x, ...)

\method{tunable}{step_measure_normalize_peak}(x, ...)

\method{tunable}{step_measure_derivative}(x, ...)

\method{tunable}{step_measure_derivative_gap}(x, ...)

\method{tunable}{step_measure_baseline_airpls}(x, ...)

\method{tunable}{step_measure_baseline_arpls}(x, ...)

\method{tunable}{step_measure_smooth_ma}(x, ...)

\method{tunable}{step_measure_smooth_median}(x, ...)

\method{tunable}{step_measure_smooth_gaussian}(x, ...)

\method{tunable}{step_measure_filter_fourier}(x, ...)

\method{tunable}{step_measure_despike}(x, ...)

\method{tunable}{step_measure_align_shift}(x, ...)

\method{tunable}{step_measure_qc_outlier}(x, ...)

\method{tunable}{step_measure_emsc}(x, ...)

\method{tunable}{step_measure_osc}(x, ...)
}
\arguments{
\item{x}{A recipe step object}

\item{...}{Not used.}
}
\value{
A tibble object.
}
\description{
These functions define what parameters \emph{can} be tuned for specific steps.
They also define the recommended objects from the \code{dials} package that can be
used to generate new parameter values and other characteristics.
}
\keyword{internal}
